Move SiteTemplatesDataService to App\Config namespace
