+++ engine/js/ng/site-templates/site-templates.reducer.js
@@ -13,7 +13,7 @@
       switch (action.type) {
         case ActionTypes.SET_STATE:
           console.log('Template settings reducer:', action);
-          return Immutable.fromJS(action.state.site_templates);
+          return Immutable.fromJS(action.state.siteTemplates);
 
         default:
           return state;
